Place all tags in a static shared enums class
