+++ Assets/Scripts/SE.cs
@@ -1,0 +1,33 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+/* Holds shared enums. */
+public static class SE {
+
+    public enum Flavor {
+        Orange,
+        Mint,
+        PeachAndPassionFruit,
+        Spearmint
+    }
+
+    public enum PathLetter {
+        A,
+        B,
+        C,
+        D
+    }
+
+    public enum PathRoute {
+        Melee,
+        Ranged
+    }
+
+    public enum PathDistance {
+        None,
+        Close,
+        Mid,
+        Far
+    }
+}
