+++ Assets/Scripts/SE.cs.meta
@@ -1,7 +1,5 @@
 fileFormatVersion: 2
-guid: 071598e7916ee6c4f91d3e1eb6c71dae
-timeCreated: 1530567779
-licenseType: Free
+guid: b0f8d784e402ecb4bbc82cb9f715258a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
